edit: args의 타입을 float으로 교체

+++ Assets/Scripts/Class/Spawner.cs
@@ -6,7 +6,7 @@ public class Spawner {
     public string id;
     public int type;
     public float[] position;
-    public int[] args;
+    public float[] args;
     public float[] time;
     public List<Command> commands;
 }
